PaddleCompletion/requirements.txt
opencv-python>=4.6.0.66
numpy>=1.23.1
tqdm>=4.64.0
h5py>=3.7.0
pandas>=1.4.3
pillow>=9.2.0
matplotlib>=3.5.2
paddlepaddle>=2.3.2
tqdm
PyYAML>=5.1
scikit-image>=0.14.0
scipy>=1.1.0
imageio==2.9.0
numba==0.53.1
easydict
munch
natsort